requirements.txt
pandas
ruff
python-dotenv
requests
psycopg[binary]